Extract `resolveSearchPaths`
