+++ src/utils/get-possible-paths.js
@@ -1,31 +1,15 @@
-import path from 'path';
-import {unquote} from './unquote';
-
-function getFilePath(searchPath, possiblePath) {
-    const [firstPart, ...restParts] = possiblePath.split(' + ');
-    const filePath = path.resolve(searchPath, unquote(firstPart));
-
-    return restParts.length > 0 ?
-        `${[`"${filePath}"`, ...restParts].join(' + ')}` : filePath;
-}
+import {resolveSearchPaths} from './resolve-search-paths';
 
 /**
  * @param {string[]} paths
  * @param {string[]} searchPaths
- * @returns {Array.<[string, array]>}
+ * @returns {Array.<[string, string[]]>}
  */
 export function getPossiblePaths(paths, searchPaths) {
     return paths.map(function(possiblePath) {
         return [
             possiblePath,
-            searchPaths.map((searchPath) => [
-                path.resolve(searchPath),
-                getFilePath(searchPath, possiblePath)
-            ]).filter(function([basePath, filePath]) {
-                return unquote(filePath).startsWith(basePath);
-            }).map(function([, filePath]) {
-                return filePath;
-            })
+            resolveSearchPaths(possiblePath, searchPaths)
         ];
     });
 }
